move tag dialogs into separate folder
